% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spect.R
\name{createSynData}
\alias{createSynData}
\title{Generates a survival data set for synthetic streaming service subscription data. The survival 
event in this case is a cancellation of the subscription. It is given as a function of 
household income and average number of hours watched in the prior month. Users can adjust the
level of censoring and variance in the data with teh supplied parameters.}
\usage{
createSynData(
  sample_size = 250,
  minimum_income = 5000,
  median_income = 50000,
  income_variance = 10000,
  min_watchhours = 0,
  max_watchhours = 6,
  censor_percentage = 0,
  min_censor_amount = 0,
  max_censor_amount = 0,
  study_time_in_months = 48,
  perturbation_shift = 0
)
}
\arguments{
\item{sample_size}{(= 250) size of the sample population to generate}

\item{minimum_income}{(= 5000) minimum household income used to generate the distribution}

\item{median_income}{(= 50000) median household income used to generate the distribution}

\item{income_variance}{(= 10000) variance to use when generating the household income distribution}

\item{min_watchhours}{(= 0.0) minimum average number of hours watched used to generate the distribution}

\item{max_watchhours}{(= 6.0) minimum average number of hours watched used to generate the distribution}

\item{censor_percentage}{(= 0.0) percentage of population to artificially censor}

\item{min_censor_amount}{(= 0.0) Minimum number of months of censoring to apply to the censored population}

\item{max_censor_amount}{(= 0.0) maximum number of months of censoring to apply to the censored population}

\item{study_time_in_months}{(= 48) observation horizon in months}

\item{perturbation_shift}{(= 0) defines a boundary for the amount to randomly perturb the formulaic result. Zero for no perturbation}
}
\value{
nothing
}
\description{
Generates a survival data set for synthetic streaming service subscription data. The survival 
event in this case is a cancellation of the subscription. It is given as a function of 
household income and average number of hours watched in the prior month. Users can adjust the
level of censoring and variance in the data with teh supplied parameters.
}
\examples{
createSynData(param)

}
\seealso{
\code{\link{createSynData}}
}
\author{
Stephen Abrams, \email{stephen.abrams@louisville.edu}
}
\keyword{utilities}
