% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spect.R
\name{evaluate_model}
\alias{evaluate_model}
\title{Generates evaluation metrics, include time-dependent TPR and FPR rates as well as AUC}
\usage{
evaluate_model(train_result, prediction_times, plot_roc = TRUE)
}
\arguments{
\item{train_result}{- return data object from spect_train}

\item{prediction_times}{- a vecotr of times to use}

\item{plot_roc}{(=TRUE) - indicates whether to display the time-dependent ROC curves.
The TPR and FPR data will be returned regardless of the value of this parameter.}
}
\value{
None - plots the number of requested samples
}
\description{
Generates evaluation metrics, include time-dependent TPR and FPR rates as well as AUC
}
\examples{
evaluate_model(param)

}
\author{
Stephen Abrams, \email{stephen.abrams@louisville.edu}
}
\keyword{utilities}
