% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spect.R
\name{plot_KM}
\alias{plot_KM}
\title{Plots a series of population Kaplan-Meier curves for different thresholds for both the test 
predictions and the ground truth}
\usage{
plot_KM(train_result, prediction_threshold_search_granularity = 0.05)
}
\arguments{
\item{train_result}{- return data object from spect_train}

\item{prediction_threshold_search_granularity}{(= 0.05) # A number between zero and one which defines the granularity of searching for cumulative probability thresholds. 
For instance, search a value of 0.05 will search 19 thresholds (0.05, 0.10, ..., 0.95)}
}
\value{
None - plots the number of requested samples
}
\description{
Plots a series of population Kaplan-Meier curves for different thresholds for both the test 
predictions and the ground truth
}
\examples{
plot_KM(param)

}
\seealso{
\code{\link{plot_KM}}
}
\author{
Stephen Abrams, \email{stephen.abrams@louisville.edu}
}
\keyword{utilities}
