% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spect.R
\name{plot_samples}
\alias{plot_samples}
\title{Plots a sample of individual survival curves from the test data set.}
\usage{
plot_samples(train_result, samples_to_plot)
}
\arguments{
\item{train_result}{- return data object from spectTrain}

\item{samples_to_plot}{- number of samples to plot}
}
\value{
None - plots the number of requested samples
}
\description{
Plots a sample of individual survival curves from the test data set.
}
\examples{
plot_samples(param)

}
\seealso{
\code{\link{plot_samples}}
}
\author{
Stephen Abrams, \email{stephen.abrams@louisville.edu}
}
\keyword{utilities}
