% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spect.R
\name{plot_survival_curve}
\alias{plot_survival_curve}
\title{Plots a sample of individual survival curves from the test data set.}
\usage{
plot_survival_curve(train_result, individual_id, curve_type = "both")
}
\arguments{
\item{train_result}{return data object from `spect_train`}

\item{individual_id}{identifier of the individual to plot}

\item{curve_type}{optional specification of the type of curve. Available options are 
"conditional", which plots the conditional probability of surviving each interval given
that the individual survived to the start of that interval, "absolute" which plots the 
unconditional probability of surviving each interval, and "both", the default value, 
which plots both curves on the same chart.}
}
\value{
None - plots the number of requested samples
}
\description{
Plots a sample of individual survival curves from the test data set.
}
\author{
Stephen Abrams, \email{stephen.abrams@louisville.edu}
}
\keyword{visualization}
