% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spect.R
\name{plot_synthetic_data}
\alias{plot_synthetic_data}
\title{Simple visualization of synthetic subscription data.}
\usage{
plot_synthetic_data(data)
}
\arguments{
\item{data}{a data object generated by create_synthetic_data}
}
\value{
None - prints synthetic data generated by create_synthetic_data
}
\description{
Simple visualization of synthetic subscription data.
}
\examples{
data <- create_synthetic_data()
plot_synthetic_data(data)

}
\author{
Stephen Abrams, \email{stephen.abrams@louisville.edu}
}
\keyword{utilities}
