% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spect.R
\name{plotSynData}
\alias{plotSynData}
\title{Simple visualization of synthetic subscription data.}
\usage{
plotSynData(data)
}
\arguments{
\item{data}{a data object generated by createSynData}
}
\value{
None - prints synthetic data generated by createSynData
}
\description{
Simple visualization of synthetic subscription data.
}
\examples{
data <- createSynData()
plotSynData(data)

}
\seealso{
\code{\link{plotSynData}}
}
\author{
Stephen Abrams, \email{stephen.abrams@louisville.edu}
}
\keyword{utilities}
