% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spect.R
\name{spect_predict}
\alias{spect_predict}
\title{Plots a series of population Kaplan-Meier curves for different thresholds for both the test 
predictions and the ground truth}
\usage{
spect_predict(train_result, new_data)
}
\arguments{
\item{train_result}{- return data object from spect_train}

\item{newdata}{- New data set with the same covariates as the training data set.}
}
\value{
predictions by the trained model on a new data set
}
\description{
Plots a series of population Kaplan-Meier curves for different thresholds for both the test 
predictions and the ground truth
}
\examples{
spect_predict(param)

}
\author{
Stephen Abrams, \email{stephen.abrams@louisville.edu}
}
\keyword{utilities}
