% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spect.R
\name{spect_train}
\alias{spect_train}
\title{Trains an ensemble model on the supplied survival data}
\usage{
spect_train(
  test_prop = 0.2,
  censor_type = "half",
  bin_slices = 10,
  method = "repeatedcv",
  resampling_number = 10,
  kfold_repeats = 3,
  model_algorithm,
  base_learner_list = list(),
  metric = "Kappa",
  rng_seed = 1101,
  use_parallel = TRUE,
  cores = 0,
  modeling_data,
  event_indicator_var,
  survival_time_var,
  obs_window
)
}
\arguments{
\item{test_prop}{(= 0.2) - proportion of the data set to reserve for testing}

\item{censor_type}{(= 'half') method used to determine censorship in a given bin - may be "half", "prev" or "same". see createDiscreteDat for usage.}

\item{bin_slices}{(= 10) Number of intervals to use for predictions.}

\item{method}{(= 'repeatedcv') caret parameter}

\item{resampling_number}{(= 10) for repeated cv}

\item{kfold_repeats}{(= 3) number of folds}

\item{model_algorithm}{Ensemble meta-learner algorithm.}

\item{base_learner_list}{# base learner algorithms}

\item{metric}{(= 'Kappa') Is there a better metric? Maybe a custom one?}

\item{rng_seed}{(= 42) Optionally set the RNG seed for reproducibility}

\item{use_parallel}{(= TRUE) make use of Caret multicore training cluster}

\item{cores}{(= 0) Only relevant if use_parallel is set to TRUE. If zero, spect will attempt to make a good choice}

\item{modeling_data}{This data set must have one column for time, one for the indicator, and the rest are treated as covariates}

\item{event_indicator_var}{The name of the column containing the event indicator (zero or one)}

\item{survival_time_var}{The name of the column containing the time variable}

\item{obs_window}{The last time to use for generating person-period data. 
Any event occurring after this time will be administratively censored. In general, 
choosing a time at or near the end of the max observed time will include most events.}

\item{flog_level}{(= INFO) Logging verbosity - options are INFO, WARN, and DEBUG. Note, this can also be set in the calling function using the fultile.logger package.}
}
\value{
Vector of character strings representing the chosen set of colors, in RGB.
}
\description{
Generates a stacked ensemble model with the given caret models as base learners
 and metalearner.
}
\examples{
spect_train(param)

}
\seealso{
\code{\link{spect_train}}
}
\author{
Stephen Abrams, \email{stephen.abrams@louisville.edu}
}
\keyword{utilities}
